Add function to extract bzr rev from git tags.
Added a function which can extract the last merged-in bzr revision from
existing tags in Git. These tags must be named like "bzr<rev>", that is,
the word "bzr" followed by the revision number. The function minds both
the tags in the .git/refs/tags folder and in the .git/packed-refs file.