Add .travis.yml file to disable CI-build
This file disables building for this branch and disables email notifications,
just in case someone forgets to disable the travis hook before pushing the
bazaar-integration branch (*cough*me*cough*).
We'll have to see how this plays along with new upstream changes, so I'll only
push it to the orga repo once I've successfully tested it in my own.